# docs/swagger.yaml
...
info:
  contact: {}
paths:
...
  /ecopedia:
    get:
      consumes:
      - application/json
      description: Dapatkan semua data Ecopedia dengan opsi ID sebagai parameter query
        opsional
      parameters:
      - description: ID Ecopedia (opsional)
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Dapatkan semua data Ecopedia
      tags:
      - Ecopedia
    post:
      consumes:
      - multipart/form-data
      description: Buat data Ecopedia baru dengan informasi yang diberikan
      parameters:
      - description: File gambar 1
        in: formData
        name: file1
        required: true
        type: file
      - description: File gambar 2
        in: formData
        name: file2
        required: true
        type: file
      - description: judul
        in: formData
        name: judul
        required: true
        type: string
      - description: sub_judul
        in: formData
        name: SubJudul
        required: true
        type: string
      - description: Deskripsi
        in: formData
        name: Deskripsi
        required: true
        type: string
      - description: src_gambar
        in: formData
        name: SrcGambar
        required: true
        type: string
      - description: referensi
        in: formData
        name: Referensi
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Buat data Ecopedia baru
      tags:
      - Ecopedia
  /ecopedia/{id}:
    delete:
      consumes:
      - application/json
      description: Hapus data Ecopedia berdasarkan ID yang diberikan
      parameters:
      - description: Ecopedia ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Hapus data Ecopedia berdasarkan ID
      tags:
      - Ecopedia
    get:
      consumes:
      - application/json
      description: Dapatkan data Ecopedia berdasarkan ID yang diberikan
      parameters:
      - description: Ecopedia ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Dapatkan data Ecopedia berdasarkan ID
      tags:
      - Ecopedia
    put:
      consumes:
      - multipart/form-data
      description: Update data Ecopedia berdasarkan ID yang diberikan dengan informasi
        yang diberikan
      parameters:
      - description: Ecopedia ID
        in: path
        name: id
        required: true
        type: integer
      - description: File gambar
        in: formData
        name: file
        required: true
        type: file
      - description: judul
        in: formData
        name: judul
        required: true
        type: string
      - description: sub_judul
        in: formData
        name: SubJudul
        required: true
        type: string
      - description: Deskripsi
        in: formData
        name: Deskripsi
        required: true
        type: string
      - description: src_gambar
        in: formData
        name: SrcGambar
        required: true
        type: string
      - description: referensi
        in: formData
        name: Referensi
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update data Ecopedia
      tags:
      - Ecopedia
  /ecopedia/comment/{id}:
    post:
      consumes:
      - application/json
      description: Tambahkan komentar atau tindakan pengguna terhadap Ecopedia berdasarkan
        ID yang diberikan
      parameters:
      - description: ID Ecopedia
        in: path
        name: id
        required: true
        type: integer
      - description: Komentar
        in: formData
        name: comment
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Tambahkan komentar atau tindakan pengguna terhadap Ecopedia
      tags:
      - Ecopedia
...
swagger: "2.0"